SQL Key Long to Int

+++ odcleanstore/webfrontend/src/main/java/cz/cuni/mff/odcleanstore/webfrontend/bo/Role.java
@@ -36,7 +36,7 @@ public class Role extends EntityWithSurrogateKey
 	 * @param label
 	 * @param description
 	 */
-	public Role(Long id, String label, String description)
+	public Role(Integer id, String label, String description)
 	{
 		super(id);
