+++ odcleanstore/webfrontend/src/main/java/cz/cuni/mff/odcleanstore/webfrontend/dao/oi/OIRulesGroupRowMapper.java
@@ -14,7 +14,7 @@ public class OIRulesGroupRowMapper extends CustomRowMapper<OIRulesGroup>
 	{
 		return new OIRulesGroup
 		(
-			rs.getLong("id"),
+			rs.getInt("id"),
 			blobToString(rs.getBlob("label")),
 			blobToString(rs.getBlob("description"))
 		);
