+++ odcleanstore/webfrontend/src/main/java/cz/cuni/mff/odcleanstore/webfrontend/pages/pipelines/EditTransformerPage.java
@@ -21,7 +21,7 @@ public class EditTransformerPage extends FrontendPage
 	
 	private DaoForEntityWithSurrogateKey<Transformer> transformerDao;
 	
-	public EditTransformerPage(final Long id) 
+	public EditTransformerPage(final Integer id) 
 	{
 		super
 		(
@@ -40,7 +40,7 @@ public class EditTransformerPage extends FrontendPage
 		addEditTransformerForm(id);
 	}
 	
-	private void addEditTransformerForm(final Long id)
+	private void addEditTransformerForm(final Integer id)
 	{
 		Transformer transformer = transformerDao.load(id);
 		IModel<Transformer> formModel = new CompoundPropertyModel<Transformer>(transformer);
