+++ odcleanstore/webfrontend/src/main/java/cz/cuni/mff/odcleanstore/webfrontend/pages/transformers/dn/DNGroupsListPage.java
@@ -121,7 +121,7 @@ public class DNGroupsListPage extends FrontendPage
 		add(new UnobtrusivePagingNavigator("navigator", dataView));
 	}
 	
-	private Link createRerunAffectedGraphsButton(final Long groupId)
+	private Link createRerunAffectedGraphsButton(final Integer groupId)
 	{
 		Link button = new Link("rerunAffectedGraphs")
 		{
